+++ src/main/java/edu/hw10/task2/Cache.java
@@ -1,0 +1,12 @@
+package edu.hw10.task2;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Retention(RetentionPolicy.RUNTIME)
+@Target(ElementType.METHOD)
+public @interface Cache {
+    boolean persist() default false;
+}
