+++ src/main/java/edu/hw10/task2/Concatenate.java
@@ -1,0 +1,9 @@
+package edu.hw10.task2;
+
+public class Concatenate implements StringConcatenation {
+
+    @Override
+    public String concatenate(String firstString, String secondString) {
+        return firstString + secondString;
+    }
+}
