+++ src/main/java/edu/hw10/task2/FibCalculator.java
@@ -1,0 +1,6 @@
+package edu.hw10.task2;
+
+public interface FibCalculator {
+    @Cache
+    long fib(int number);
+}
