+++ src/main/java/edu/hw10/task2/Square.java
@@ -1,0 +1,9 @@
+package edu.hw10.task2;
+
+public class Square implements SquareOfNumber {
+
+    @Override
+    public int square(int number) {
+        return number * number;
+    }
+}
