+++ src/main/java/edu/hw10/task2/SquareOfNumber.java
@@ -1,0 +1,6 @@
+package edu.hw10.task2;
+
+public interface SquareOfNumber {
+    @Cache
+    int square(int number);
+}
