+++ src/main/java/edu/hw10/task2/StringConcatenation.java
@@ -1,0 +1,6 @@
+package edu.hw10.task2;
+
+public interface StringConcatenation {
+    @Cache
+    String concatenate(String firstString, String secondString);
+}
